패턴 세부속성 지정 하는중.

+++ Classes/IntSeries.h
@@ -141,7 +141,7 @@ public:
 		x = t_x;
 		y = t_y;
 	}
-	int length() const
+	float length() const
 	{
 		return sqrt(x*x + y*y);
 	}
